+++ scripts/eslint/index.js
@@ -1,3 +1,6 @@
 const commentCheck = require("./comment-check");
+const importCheck = require("./import-check");
 
-module.exports = { rules: { "comment-check": commentCheck } };
+module.exports = {
+  rules: { "comment-check": commentCheck, "import-check": importCheck },
+};